CylinderからExpand系を削除して、`rough_*`系に切り替え
