+++ src/geometry/shapes/cylinder/impls.rs
@@ -1,7 +1,4 @@
-use crate::{
-    Coordinate, CoverRangeIds, CoverSingleIds, Cylinder, Error, ExpandSolids, RangeId, Shape,
-    SingleId,
-};
+use crate::{Coordinate, CoverRangeIds, CoverSingleIds, Cylinder, Error, RangeId, Shape, SingleId};
 
 impl Shape for Cylinder {
     fn center(&self) -> Coordinate {
@@ -11,7 +8,7 @@ impl Shape for Cylinder {
 
 impl CoverSingleIds for Cylinder {
     fn cover_single_ids(&self, z: u8) -> Result<impl Iterator<Item = SingleId>, Error> {
-        let solid = self.expand_solids().next().unwrap();
+        let solid = self.rough_solid();
         let ids: Vec<_> = solid.cover_single_ids(z)?.collect();
         Ok(ids.into_iter())
     }
@@ -19,7 +16,7 @@ impl CoverSingleIds for Cylinder {
 
 impl CoverRangeIds for Cylinder {
     fn cover_range_ids(&self, z: u8) -> Result<impl Iterator<Item = RangeId>, Error> {
-        let solid = self.expand_solids().next().unwrap();
+        let solid = self.rough_solid();
         let ids: Vec<_> = solid.cover_range_ids(z)?.collect();
         Ok(ids.into_iter())
     }
