+++ src/geometry/shapes/cylinder/mod.rs
@@ -1,14 +1,14 @@
-pub mod geometry_relation;
 pub mod impls;
+use std::f64::consts::PI;
 
-use crate::{Coordinate, Error, GeometryError};
+use crate::{Coordinate, Ecef, Error, GeometryError, Polygon, Solid, Vec3};
 #[derive(Debug, Clone, Copy, PartialEq)]
 /// 3次元空間における円柱を表す型。
 ///
 /// 中心線及び半径によって定義される立体的な領域を表現する。
 pub struct Cylinder {
     pub start: Coordinate,
-    pub end: Coordinate, //向きに明確な意味を持たせるならstart,endを別に定義する方がいい？
+    pub end: Coordinate,
     pub radius_m: f64,
 }
 
@@ -23,5 +23,61 @@ impl Cylinder {
         } else {
             Err(GeometryError::RadiusNegative { radius: radius_m }.into())
         }
+    }
+
+    pub fn rough_solid(&self) -> Solid {
+        let polygons = self.rough_surfaces().collect();
+        Solid::new(polygons, 1e-10).unwrap()
+    }
+
+    pub fn rough_surfaces(&self) -> impl Iterator<Item = Polygon> {
+        let vecs: [Vec3; 2] = [self.start.into(), self.end.into()];
+        let vec_n = vecs[1] - vecs[0];
+        let basis = vec_n
+            .create_orthonormal_basis()
+            .map(|v| v.scale(self.radius_m));
+        let divide_num = 100_u32;
+        let vertices: Vec<_> = (0..divide_num)
+            .map(|i| {
+                let theta = 2.0 * PI * i as f64 / divide_num as f64;
+                let v = basis[0].scale(theta.cos()) + basis[1].scale(theta.sin());
+                [vecs[0] + v, vecs[1] + v]
+            })
+            .collect();
+
+        let to_coord = |v: Vec3| Coordinate::try_from(Ecef::from(v)).unwrap();
+
+        // 側面の頂点リスト（イテレータ）を作成
+        let side_surfaces = (0..divide_num).map(|i| {
+            let next_i = (i + 1) % divide_num;
+            let v_curr = vertices[i as usize];
+            let v_next = vertices[next_i as usize];
+
+            Polygon::new(
+                vec![
+                    to_coord(v_curr[0]),
+                    to_coord(v_next[0]),
+                    to_coord(v_next[1]),
+                    to_coord(v_curr[1]),
+                ],
+                1e-10,
+            )
+        });
+
+        // 側面を全て集める
+        let mut raw_surfaces: Vec<Polygon> = side_surfaces.collect();
+
+        // 底面
+        raw_surfaces.push(Polygon::new(
+            vertices.iter().rev().map(|v| to_coord(v[0])).collect(),
+            1e-10,
+        ));
+        // 上面
+        raw_surfaces.push(Polygon::new(
+            vertices.iter().map(|v| to_coord(v[1])).collect(),
+            1e-10,
+        ));
+
+        raw_surfaces.into_iter()
     }
 }
